Rename CameraView into CaptureTour
