+++ packages/react-native-views/src/components/CaptureTour/CameraControls/index.js
