+++ packages/react-native-views/src/components/CaptureTour/CameraOrientationView/drawing.js
