+++ packages/react-native-views/src/components/CaptureTour/CameraOrientationView/index.js
