+++ packages/react-native-views/src/components/CaptureTour/CameraOverlay/index.js
