+++ packages/react-native-views/src/components/CaptureTour/CameraScrollView/index.js
