+++ packages/react-native-views/src/components/CaptureTour/CameraSimpleView/index.js
