+++ packages/react-native-views/src/components/CaptureTour/hooks/useMobileBrowserConfig.js
