+++ packages/react-native-views/src/components/CaptureTour/hooks/usePictures.js
