+++ packages/react-native-views/src/components/CaptureTour/hooks/useSights.js
