+++ packages/react-native-views/src/components/index.js
@@ -1,6 +1,8 @@
 export { default as ActivityIndicatorView } from './ActivityIndicatorView';
 export { default as AdvicesView } from './AdvicesView';
-export { default as CameraView } from './CameraView';
+export { default as CameraView } from './CaptureTour'; // Deprecated name
+export { default as CameraSimpleView } from './CaptureTour/CameraSimpleView';
+export { default as CaptureTour } from './CaptureTour';
 export { default as DamagedPartsView } from './DamagedPartsView';
 export { default as DamagesView } from './Damages';
 export { default as PicturesSummaryView } from './PicturesSummaryView';
